# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "這是我們的 API 文件",
        "contact": {
            "email": "xxx"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/createcurrency": {
            "post": {
                "tags": [
                    "base_currency"
                ],
                "summary": "新增貨幣資訊",
                "description": "新增貨幣資訊",
                "operationId": "createcurrency",
                "parameters": [
                    {
                        "name": "currency_no",
                        "in": "query",
                        "description": "貨幣代號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "currency_nm",
                        "in": "query",
                        "description": "貨幣名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "currency_rate",
                        "in": "query",
                        "description": "現在匯率(以台幣為基準)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Note",
                        "in": "query",
                        "description": "備註",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_valid",
                        "in": "query",
                        "description": "是否有效",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "currency_no": {
                                            "type": "string",
                                            "example": "C001"
                                        },
                                        "currency_nm": {
                                            "type": "string",
                                            "example": "台幣"
                                        },
                                        "currency_rate": {
                                            "type": "string",
                                            "example": "取得當下匯率"
                                        },
                                        "note": {
                                            "type": "string",
                                            "example": "測試測試"
                                        },
                                        "is_valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "create_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "create_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "建立失敗"
                    }
                }
            }
        },
        "/api/currency/{currencyno}": {
            "get": {
                "tags": [
                    "base_currency"
                ],
                "summary": "查詢特定貨幣資訊",
                "description": "查詢特定貨幣資訊",
                "operationId": "getcurrency",
                "parameters": [
                    {
                        "name": "currencyno",
                        "in": "path",
                        "description": "貨幣代號",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "CurrencyNo": {
                                            "type": "string",
                                            "example": "C001"
                                        },
                                        "CurrencyNM": {
                                            "type": "string",
                                            "example": "台幣"
                                        },
                                        "Note": {
                                            "type": "string",
                                            "example": "測試測試"
                                        },
                                        "is_valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "Createuser": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "CreateTime": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到貨幣資訊"
                    }
                }
            }
        },
        "/api/currency2/{currencynm}": {
            "get": {
                "tags": [
                    "base_currency"
                ],
                "summary": "查詢特定貨幣資訊",
                "description": "查詢特定貨幣資訊",
                "operationId": "getcurrencynm",
                "parameters": [
                    {
                        "name": "currencynm",
                        "in": "path",
                        "description": "貨幣名稱",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "CurrencyNo": {
                                            "type": "string",
                                            "example": "C001"
                                        },
                                        "CurrencyNM": {
                                            "type": "string",
                                            "example": "台幣"
                                        },
                                        "Note": {
                                            "type": "string",
                                            "example": "測試測試"
                                        },
                                        "is_valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "Createuser": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "CreateTime": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到貨幣資訊"
                    }
                }
            }
        },
        "/api/currencys/valid": {
            "get": {
                "tags": [
                    "base_currency"
                ],
                "summary": "查詢所有有效貨幣資訊",
                "description": "查詢所有有效貨幣資訊",
                "operationId": "getallcurrency",
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "example": "0b422f02-5acf-4bbb-bddf-4f6fdd843b08"
                                        },
                                        "CurrencyNo": {
                                            "type": "string",
                                            "example": "C001"
                                        },
                                        "CurrencyNM": {
                                            "type": "string",
                                            "example": "台幣"
                                        },
                                        "Note": {
                                            "type": "string",
                                            "example": "測試測試"
                                        },
                                        "is_valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "Createuser": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "update_user": {
                                            "type": "string",
                                            "example": "admin"
                                        },
                                        "CreateTime": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001975Z"
                                        },
                                        "update_time": {
                                            "type": "string",
                                            "example": "2025-03-31T08:58:52.001986Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "未找到有效貨幣資訊"
                    }
                }
            }
        },
...
        "/api/currencys/showconst": {
            "get": {
                "tags": [
                    "base_currency"
                ],
                "summary": "列出所有幣別需要的常用(下拉、彈窗)",
                "description": "列出所有幣別需要的常用(下拉、彈窗)",
                "operationId": "Show_currency_aLL_const",
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "404": {
                        "description": "有效單據類型未找到"
                    }
                }
            }
        },
...
    },
...
}